rename cons_gr to gr
