+++ packages/chain/cons/gr/gr.go
@@ -3,7 +3,7 @@
 
 // The purpose of this package is to run the consensus protocol
 // as a goroutine and communicate with all the related components.
-package cons_gr
+package gr
 
 import (
 	"context"
