+++ packages/chain/cons/gr/vm_async.go
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-package cons_gr
+package gr
 
 import (
 	"context"
